Minor: gets '/signup/page.tsx' ready for development
